refactor: #118 - 유틸 함수 이동
